requirements.txt
asgiref==3.3.1
astroid==2.5.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.9.0
flake8-isort==4.0.0
importlib-metadata==3.7.3
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.1
